requirements.txt
fastapi
uvicorn
pymysql
pydantic
python-multipart
longbridge
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
